refactor: move navMap to utils
