+++ src/untils/NavMap/NavMap.ts
@@ -1,7 +1,7 @@
 class NavMap {
   constructor (private baseUrl = '') {}
 
-  home = () => `${this.baseUrl}`;
+  home = () => `${this.baseUrl}/`;
   lobby = () => `${this.baseUrl}/lobby/:id`;
   game = () => `${this.baseUrl}/game/:id`;
 };
